Refactorization and ordering
