+++ generator/chart.go
@@ -2,9 +2,12 @@ package generator
 
 import (
 	"fmt"
+	"log"
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/compose-spec/compose-go/types"
 
 	"katenary/generator/labelStructs"
 	"katenary/utils"
@@ -112,5 +115,215 @@ func (chart *HelmChart) SaveTemplates(templateDir string) {
 
 		f.Write(t)
 		f.Close()
+	}
+}
+
+// generateConfigMapsAndSecrets creates the configmaps and secrets from the environment variables.
+func (chart *HelmChart) generateConfigMapsAndSecrets(project *types.Project) error {
+	appName := chart.Name
+	for _, s := range project.Services {
+		if s.Environment == nil || len(s.Environment) == 0 {
+			continue
+		}
+
+		originalEnv := types.MappingWithEquals{}
+		secretsVar := types.MappingWithEquals{}
+
+		// copy env to originalEnv
+		for k, v := range s.Environment {
+			originalEnv[k] = v
+		}
+
+		if v, ok := s.Labels[LabelSecrets]; ok {
+			list, err := labelStructs.SecretsFrom(v)
+			if err != nil {
+				log.Fatal("error unmarshaling secrets label:", err)
+			}
+			for _, secret := range list {
+				if secret == "" {
+					continue
+				}
+				if _, ok := s.Environment[secret]; !ok {
+					fmt.Printf("%s secret %s not found in environment", utils.IconWarning, secret)
+					continue
+				}
+				secretsVar[secret] = s.Environment[secret]
+			}
+		}
+
+		if len(secretsVar) > 0 {
+			s.Environment = secretsVar
+			sec := NewSecret(s, appName)
+			y, _ := sec.Yaml()
+			name := sec.service.Name
+			chart.Templates[name+".secret.yaml"] = &ChartTemplate{
+				Content:     y,
+				Servicename: s.Name,
+			}
+		}
+
+		// remove secrets from env
+		s.Environment = originalEnv // back to original
+		for k := range secretsVar {
+			delete(s.Environment, k)
+		}
+		if len(s.Environment) > 0 {
+			cm := NewConfigMap(s, appName)
+			y, _ := cm.Yaml()
+			name := cm.service.Name
+			chart.Templates[name+".configmap.yaml"] = &ChartTemplate{
+				Content:     y,
+				Servicename: s.Name,
+			}
+		}
+	}
+	return nil
+}
+
+func (chart *HelmChart) generateDeployment(service types.ServiceConfig, deployments map[string]*Deployment, services map[string]*Service, podToMerge map[string]*types.ServiceConfig, appName string) error {
+	// check the "ports" label from container and add it to the service
+	if err := fixPorts(&service); err != nil {
+		return err
+	}
+
+	// isgnored service
+	if isIgnored(service) {
+		fmt.Printf("%s Ignoring service %s\n", utils.IconInfo, service.Name)
+		return nil
+	}
+
+	// helm dependency
+	if isHelmDependency, err := chart.setDependencies(service); err != nil {
+		return err
+	} else if isHelmDependency {
+		return nil
+	}
+
+	// create all deployments
+	d := NewDeployment(service, chart)
+	deployments[service.Name] = d
+
+	// generate the cronjob if needed
+	chart.setCronJob(service, appName)
+
+	// get the same-pod label if exists, add it to the list.
+	// We later will copy some parts to the target deployment and remove this one.
+	if samePod, ok := service.Labels[LabelSamePod]; ok && samePod != "" {
+		podToMerge[samePod] = &service
+	}
+
+	// create the needed service for the container port
+	if len(service.Ports) > 0 {
+		s := NewService(service, appName)
+		services[service.Name] = s
+	}
+
+	// create all ingresses
+	if ingress := d.AddIngress(service, appName); ingress != nil {
+		y, _ := ingress.Yaml()
+		chart.Templates[ingress.Filename()] = &ChartTemplate{
+			Content:     y,
+			Servicename: service.Name,
+		}
+	}
+
+	return nil
+}
+
+// setChartVersion sets the chart version from the service image tag.
+func (chart *HelmChart) setChartVersion(service types.ServiceConfig) {
+	if chart.Version == "" {
+		image := service.Image
+		parts := strings.Split(image, ":")
+		if len(parts) > 1 {
+			chart.AppVersion = parts[1]
+		} else {
+			chart.AppVersion = "0.1.0"
+		}
+	}
+}
+
+// setCronJob creates a cronjob from the service labels.
+func (chart *HelmChart) setCronJob(service types.ServiceConfig, appName string) *CronJob {
+	if _, ok := service.Labels[LabelCronJob]; !ok {
+		return nil
+	}
+	cronjob, rbac := NewCronJob(service, chart, appName)
+	y, _ := cronjob.Yaml()
+	chart.Templates[cronjob.Filename()] = &ChartTemplate{
+		Content:     y,
+		Servicename: service.Name,
+	}
+
+	if rbac != nil {
+		y, _ := rbac.RoleBinding.Yaml()
+		chart.Templates[rbac.RoleBinding.Filename()] = &ChartTemplate{
+			Content:     y,
+			Servicename: service.Name,
+		}
+		y, _ = rbac.Role.Yaml()
+		chart.Templates[rbac.Role.Filename()] = &ChartTemplate{
+			Content:     y,
+			Servicename: service.Name,
+		}
+		y, _ = rbac.ServiceAccount.Yaml()
+		chart.Templates[rbac.ServiceAccount.Filename()] = &ChartTemplate{
+			Content:     y,
+			Servicename: service.Name,
+		}
+	}
+
+	return cronjob
+}
+
+// setDependencies sets the dependencies from the service labels.
+func (chart *HelmChart) setDependencies(service types.ServiceConfig) (bool, error) {
+	// helm dependency
+	if v, ok := service.Labels[LabelDependencies]; ok {
+		d, err := labelStructs.DependenciesFrom(v)
+		if err != nil {
+			return false, err
+		}
+
+		for _, dep := range d {
+			fmt.Printf("%s Adding dependency to %s\n", utils.IconDependency, dep.Name)
+			chart.Dependencies = append(chart.Dependencies, dep)
+			name := dep.Name
+			if dep.Alias != "" {
+				name = dep.Alias
+			}
+			// add the dependency env vars to the values.yaml
+			chart.Values[name] = dep.Values
+		}
+
+		return true, nil
+	}
+	return false, nil
+}
+
+// setSharedConf sets the shared configmap to the service.
+func (chart *HelmChart) setSharedConf(service types.ServiceConfig, deployments map[string]*Deployment) {
+	// if the service has the "shared-conf" label, we need to add the configmap
+	// to the chart and add the env vars to the service
+	if _, ok := service.Labels[LabelEnvFrom]; !ok {
+		return
+	}
+	fromservices, err := labelStructs.EnvFromFrom(service.Labels[LabelEnvFrom])
+	if err != nil {
+		log.Fatal("error unmarshaling env-from label:", err)
+	}
+	// find the configmap in the chart templates
+	for _, fromservice := range fromservices {
+		if _, ok := chart.Templates[fromservice+".configmap.yaml"]; !ok {
+			log.Printf("configmap %s not found in chart templates", fromservice)
+			continue
+		}
+		// find the corresponding target deployment
+		target := findDeployment(service.Name, deployments)
+		if target == nil {
+			continue
+		}
+		// add the configmap to the service
+		addConfigMapToService(service.Name, fromservice, chart.Name, target)
 	}
 }
